requirements.txt
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
isodate==0.7.2
lxml==5.3.0
platformdirs==4.3.6
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
urllib3==2.2.3
zeep==4.3.1
